% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_version}
\alias{add_version}
\title{Add timestamp and sha string to a file name}
\usage{
add_version(filename, extension = "", sha_nchar = 7, sep = "__")
}
\arguments{
\item{filename}{Path sans extension of the file to version}

\item{extension}{Extension of the file}

\item{sha_nchar}{Number of characters from the SHA to use as the version
identifier}

\item{sep}{Characters separating the version identifier from the file name}
}
\value{
A character string with the file name and the version identifier
}
\description{
An alternative to version data is to name it using the sha (unique
identifier) of the code using to generate or process the data and the time at
which the data was generated or processed. This function adds this
information, a version identifier, to a file name (character string)
}
\details{
The SHA information is retrieved using \link[git2r:sha]{git2r::sha}. If the code is not
running in a context aware of a git repository (for example when code is
running inside a container) then this function attempts to get the sha from
the environment variable \code{GITHUB_SHA}. If both of these methods fail, no sha
versioning is added.
}
\examples{
if (git2r::in_repository()) {
  add_version("my_file", "csv")
}
}
