% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airtable.R
\name{air_get_records}
\alias{air_get_records}
\title{Get records from an Airtable table}
\usage{
air_get_records(
  table,
  base_id,
  api_key = Sys.getenv("AIRTABLE_KEY"),
  query = list()
)
}
\arguments{
\item{table}{name of the table (string)}

\item{base_id}{id of the Airtable base where the table is located}

\item{api_key}{API key for Airtable. By default it looks at the
\code{AIRTABLE_KEY} environment variable.}

\item{query}{additional parameters for the airtable query like for example
fields, view, pageSize, maxRecords, sort, filterByFormula, cellFormat,
timeZone, userLocale, etc... See the airtable documentation for more
details.}
}
\value{
a list with a results field containing all records
}
\description{
Downloads an Airtable table into a list of records
}
\details{
Airtable API returns one page of records at a time. Each page will contain
\code{pageSize} (a query parameter) records, which is 100 by default. Internally,
this function will automatically download all records until \code{maxRecords} has
been reached.
}
\examples{
\dontrun{
# Get all records for the table boats
air_get_records(table = "boats", base_id = "appjEVaN8kBNXAWak")

# Get records only in a specified view
air_get_records("boats", "appjEVaN8kBNXAWak",
  query = list(view = "Boats with wrong registration")
)

# Get only the first 5 records
air_get_records("boats", "appjEVaN8kBNXAWak", query = list(maxRecords = 5))
}

}
\seealso{
air_records_to_tibble
}
