% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airtable.R
\name{air_records_to_tibble}
\alias{air_records_to_tibble}
\title{Format Airtable records to a data frame (tibble)}
\usage{
air_records_to_tibble(records)
}
\arguments{
\item{records}{List of records from an airtable table}
}
\value{
a tibble
}
\description{
Format Airtable records to a data frame (tibble)
}
