% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airtable.R
\name{air_upload_records}
\alias{air_upload_records}
\title{Upload and create Airtable records to a base}
\usage{
air_upload_records(
  body,
  table,
  base_id,
  api_key = Sys.getenv("AIRTABLE_KEY"),
  request_type = c("create", "update")
)
}
\arguments{
\item{body}{list with the records, as formatted by \code{tibble_to_air_records()}}

\item{table}{name of the table (string)}

\item{base_id}{id of the Airtable base where the table is located}

\item{api_key}{API key for Airtable. By default it looks at the
\code{AIRTABLE_KEY} environment variable.}

\item{request_type}{Whether to create a new record or update an existing one.
If updating an existing one, an \code{id_field} needs to be specified when
calling \code{tibble_to_air_records()}}
}
\value{
List of request responses
}
\description{
Takes a list created by \code{tibble_to_air_records()} and uploads it to Airtable.
If \code{request_type} is set to "update" the body list must have an id field for
each record. Each element of body must be up to 10 records long.
}
