% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess-pds-trips.R
\name{associate_pds_trips}
\alias{associate_pds_trips}
\title{Label short interval trips}
\usage{
associate_pds_trips(pds_trips = NULL)
}
\arguments{
\item{pds_trips}{The table of pds trips.}
}
\value{
A dataframe with pds trips with a new column \code{associated_to} indicating
the trips to be potentially merged.
}
\description{
The pds data are quite patchy and sometimes a single trip is recorded as 2
different trips. This function calculates the time interval between the end
and the start of a trip for each boat and labels the trips that start less
than 35 minutes from the other.
}
