% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google-drive.R
\name{authenticate_google_drive}
\alias{authenticate_google_drive}
\title{Authenticate to Google Drive}
\usage{
authenticate_google_drive(options = list(method = "service_account_key"))
}
\arguments{
\item{options}{a list with information about the authentication method.
Either "service_account_key" for authenticating an app or "OAuth" for
authenticating a real person (for example when running the code locally).}
}
\description{
Authenticate to Google Drive
}
