% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-taxa.R
\name{calculate_weights}
\alias{calculate_weights}
\title{Add weight of species to merged landings}
\usage{
calculate_weights(log_threshold = logger::DEBUG)
}
\arguments{
\item{log_threshold}{The logging threshold level}
}
\value{
No outputs. This function is used for it's side effects.
}
\description{
Downloads merged landings and calculates the weight of the catch.
}
\details{
The  file is then uploaded to the cloud. The name is the same as for merged
landings but with "_weight" at the end. The parameters needed are:

\if{html}{\out{<div class="sourceCode">}}\preformatted{surveys:
  merged_landings:
    file_prefix:
    version:
storage:
  storage_name:
    key:
    options:
      project:
      bucket:
      service_account_key:
}\if{html}{\out{</div>}}

Progress through the function is tracked using the package \emph{logger}.
}
\keyword{workflow}
