% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-fishery.R
\name{estimate_fishery_indicators}
\alias{estimate_fishery_indicators}
\title{Estimate fisheries indicators}
\usage{
estimate_fishery_indicators(log_threshold = logger::DEBUG)
}
\arguments{
\item{log_threshold}{The logging threshold level}
}
\description{
Uses the trip data to estimate various fisheries statistics.
}
\details{
The parameters needed in \code{conf.yml} are:

\if{html}{\out{<div class="sourceCode">}}\preformatted{models:
  file_prefix:
storage:
  storage_name:
    key:
    options:
      project:
      bucket:
      service_account_key:
}\if{html}{\out{</div>}}
}
\keyword{workflow}
