% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-indicators.R
\name{estimate_indicators}
\alias{estimate_indicators}
\title{Estimate fishery indicators}
\usage{
estimate_indicators(trips = NULL, vessels_metadata = NULL, region = NULL)
}
\arguments{
\item{trips}{Validated dataframe containing Timor trips data.}

\item{vessels_metadata}{Vessels metadata.}

\item{region}{Timor municipality.}
}
\value{
A list of fishery indicators divided by municipality.
}
\description{
Estimate municipal fishery indicators based on vessels registration records.
}
