% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-weights.R
\name{estimate_weight}
\alias{estimate_weight}
\title{Estimate weight of a single taxon type}
\usage{
estimate_weight(length, length_type, code, n_individuals, lw, ll)
}
\description{
Estimate weight of a single taxon type
}
