% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataverse.R
\name{export_dataverse}
\alias{export_dataverse}
\title{Export file to Dataverse}
\usage{
export_dataverse(
  token,
  dataverse,
  upfile,
  metadat,
  server = "dataverse.harvard.edu"
)
}
\arguments{
\item{token}{API token associated to Dataverse account.}

\item{dataverse}{A character string specifying a Dataverse name.}

\item{upfile}{File to be exported.}

\item{metadat}{A list of metadata in Dublin Core format, see
\url{https://www.dublincore.org/specifications/dublin-core/dcmi-terms/}.}

\item{server}{A character string specifying a Dataverse server.}
}
\description{
This function exports a file to a specified Dataverse repository.
}
\examples{
\dontrun{
metadat <- list(
  title = paste("My dataset", format(Sys.time(), "\%Y\%m\%d_\%H\%M")),
  creator = "A big dinosaur",
  description = "This is an amazing dataset...",
  language = "English"
)

upl_files <- "/Users/..../my_data.csv"

export_dataverse(
  token = "a3k2XXXX-XXXX-XXXX-XXXX-XXXXXXXXXXXX",
  dataverse = "my_dataverse",
  upfile = file,
  metadat = metadat
)
}

}
