% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-fishery.R
\name{fill_missing_regions}
\alias{fill_missing_regions}
\title{Fill missing regions}
\usage{
fill_missing_regions(trips = NULL)
}
\arguments{
\item{trips}{Dataframe  with Timor validated trips.}
}
\value{
A dataframe with filled regions (where possible).
}
\description{
Replace empty region data based on boats geographic activity (trough tracker imeis)
}
\examples{
\dontrun{
fill_missing_regions(trips)
}
}
\keyword{helper}
