% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-public-data.R
\name{format_public_data}
\alias{format_public_data}
\title{Format public data}
\usage{
format_public_data(log_threshold = logger::DEBUG)
}
\arguments{
\item{log_threshold}{The (standard Apache logj4) log level used as a
threshold for the logging infrastructure. See \link[logger:log_levels]{logger::log_levels} for more
details}
}
\value{
no outputs. This function is used for it's side effects
}
\description{
Format the merged trips into files that are fit for distribution.
Specifically, this function produces a trips and a catch table as well as
tables containing aggregated data. These files are stored both as tsv files
and rds in a storage bucket with public access.
}
\details{
#'The parameters needed in the config file are those required for
\code{merge_trips()} in addition to:

\if{html}{\out{<div class="sourceCode">}}\preformatted{public_storage:
 google:
   key: gcs
   options:
     project:
     bucket:
     service_account_key:
export:
 file_prefix:
}\if{html}{\out{</div>}}
}
\keyword{workflow}
