% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-indicators.R
\name{generate_estimates}
\alias{generate_estimates}
\title{Model fisheries indicators}
\usage{
generate_estimates(log_threshold = logger::DEBUG)
}
\arguments{
\item{log_threshold}{The (standard Apache logj4) log level used as a
threshold for the logging infrastructure. See \link[logger:log_levels]{logger::log_levels} for more
details}
}
\description{
Uses the trip data to model various fisheries statistics.
}
\details{
The parameters needed in \code{conf.yml} are:

\if{html}{\out{<div class="sourceCode">}}\preformatted{models:
  file_prefix:
storage:
  storage_name:
    key:
    options:
      project:
      bucket:
      service_account_key:
}\if{html}{\out{</div>}}
}
\keyword{workflow}
