% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-dataverse.R
\name{generate_metadata}
\alias{generate_metadata}
\title{Generate a list of metadata}
\usage{
generate_metadata(conf, temp_coverage = NULL)
}
\arguments{
\item{conf}{The configuration file.}

\item{temp_coverage}{Temporal coverage of the data to upload.}
}
\value{
A list with metadata information
}
\description{
The function generate a list of metadata information to append to the
files to upload to a Dataverse repository.
}
\examples{
\dontrun{
conf <- read_config()
generate_metadata(conf, temp_coverage = "2018-2024")
}
}
\keyword{export}
