% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length-params.R
\name{get_catch_types}
\alias{get_catch_types}
\title{Download metadata catch types}
\usage{
get_catch_types(log_threshold = logger::DEBUG)
}
\arguments{
\item{log_threshold}{The (standard Apache logj4) log level used as a
threshold for the logging infrastructure. See \link[logger:log_levels]{logger::log_levels} for more
details}
}
\value{
No output. This funcrion is used for it's side effects
}
\description{
This function download airtable tables containing metadata of catch types.
}
