% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve-survey-data.R
\name{get_host_url}
\alias{get_host_url}
\title{Select server from kobo}
\usage{
get_host_url(api, version = "v1")
}
\arguments{
\item{api}{Either "kobo", "kobohr", "ona", or a custom (full) URL. API URLs
are made available for KoBo Toolbox ("kobo",
\url{https://kc.kobotoolbox.org/api/v1/}), KoBo Humanitarian Response
("kobohr", \url{https://kc.humanitarianresponse.info/api/v1/}), Ona ("ona",
\url{https://api.ona.io/api/v1/}) and Unhcr ("unhcr",
\url{https://kobocat.unhcr.org/api/v1/}) . For your own installation, or
other installations using the same API but accessed at a different URL,
enter the full URL.}

\item{version}{Wether using the API "v1" or "v2"}
}
\value{
An URL
}
\description{
Specifies the Host URL of the API to use
}
\examples{

get_host_url("unhcr")
get_host_url("https://kobocat.unhcr.org/api/v1/")
}
\author{
Ananda Mahto
}
