% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-dataverse.R
\name{get_metadata_info}
\alias{get_metadata_info}
\title{Retrieve metadata information from survey landings}
\usage{
get_metadata_info(pars)
}
\arguments{
\item{pars}{Configuration file}
}
\value{
A list with metadata information:
extent: time period covered
....
}
\description{
This function downloads the latest version of landings and extract metadata
information.
}
