% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-cloud-files.R
\name{get_models}
\alias{get_models}
\title{Download models estimates}
\usage{
get_models(conf)
}
\arguments{
\item{conf}{Configuration file}
}
\value{
A list of dataframes with national and municipal estimations
}
\description{
Download models estimates
}
\keyword{storage}
