% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve-pds-data.R
\name{get_pds_resp}
\alias{get_pds_resp}
\title{Convert API response to data frame}
\usage{
get_pds_resp(data, secret = NULL, token = NULL, id = NULL)
}
\arguments{
\item{data}{A character indicating either "trips" or "tracks"}

\item{secret}{Access secret code for the account}

\item{token}{Access token for the account}

\item{id}{Trip unique id from which retrieve tracks data}
}
\value{
A data frame with tracks or trips data
}
\description{
This function get Pelagic Data System API responses and return them as a
data frame. It is used inside ingest functions in order to process multiple
API responses.
}
\examples{
\dontrun{

  get trips data:
  get_pds_resp("trips",secret = "abcXXXXXX", token = "123XXXXXX")

  get tracks data:
  get_pds_resp("tracks",secret = "abcXXXXXX",token = "123XXXXXX", id="2327722")
}

}
