% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-indicators.R
\name{get_vessels_records}
\alias{get_vessels_records}
\title{Get updated vessels records}
\usage{
get_vessels_records(table = NULL, base_id = NULL, api_key = NULL)
}
\arguments{
\item{table}{name of the table (string)}

\item{base_id}{id of the Airtable base where the table is located}

\item{api_key}{API key for Airtable. By default it looks at the
\code{AIRTABLE_KEY} environment variable.}
}
\value{
A dataframe.
}
\description{
Get 2016 and ongoing info on the number of vessels recorded from Timor MAF.
}
