% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pds-maps.R
\name{ingest_pds_map}
\alias{ingest_pds_map}
\title{Generate and ingest Timor maps}
\usage{
ingest_pds_map(log_threshold = logger::DEBUG)
}
\arguments{
\item{log_threshold}{The (standard Apache logj4) log level used as a
threshold for the logging infrastructure. See \link[logger:log_levels]{logger::log_levels} for more
details}
}
\value{
No output. This function is used for it's side effects.
}
\description{
This function downloads pds tracks coordinates, generates a png image
showing the map of Timor divided by municipalities including the tracks paths,
and upload it to cloud storage. It also upload the data frame splitted by grids
to produce leaflet maps in the portal.
}
