% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge-trips.R
\name{ingest_pds_matched_trips_TEST}
\alias{ingest_pds_matched_trips_TEST}
\title{Test ingestion of zip folder}
\usage{
ingest_pds_matched_trips_TEST(log_threshold = logger::DEBUG)
}
\arguments{
\item{log_threshold}{The (standard Apache logj4) log level used as a
threshold for the logging infrastructure. See \link[logger:log_levels]{logger::log_levels} for more
details}
}
\value{
no outputs. This function is used for it's side effects
}
\description{
Ingest a dummy zip folder to cloud storage.
}
