% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest-pds-data.R
\name{ingest_pds_tracks}
\alias{ingest_pds_tracks}
\title{Ingest Pelagic Data System tracks data}
\usage{
ingest_pds_tracks(log_threshold = logger::DEBUG)
}
\arguments{
\item{log_threshold}{The (standard Apache logj4) log level used as a
threshold for the logging infrastructure. See \link[logger:log_levels]{logger::log_levels} for more
details}
}
\value{
No output. This funcrion is used for it's side effects
}
\description{
This function read Pelagic Data System (pds) tracks data, merge them with
trips information by their id, then uploads it to cloud storage services as
a csv data frame.
}
\details{
The function  downloads and uploads only the tracks data that are not yet
stored in the bucket.

The parameters needed in \code{conf.yml} are:\preformatted{pds:
  tracks:
    token:
    secret:
    file_prefix:
pds_storage:
  storage_name:
    key:
    options:
      project:
      bucket:
      service_account_key:
}

Progress through the function is tracked using the package \emph{logger}.
}
\keyword{workflow}
