% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-weights.R
\name{ingest_rfish_table}
\alias{ingest_rfish_table}
\title{Ingest species weight-length info}
\usage{
ingest_rfish_table(log_threshold = logger::DEBUG)
}
\arguments{
\item{log_threshold}{The (standard Apache logj4) log level used as a
threshold for the logging infrastructure. See \link[logger:log_levels]{logger::log_levels} for more
details}
}
\value{
No output. This function is used for it's side effects
}
\description{
Ingest species weight-length info
}
