% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest-landings.R
\name{ingest_timor_landings}
\alias{ingest_timor_landings}
\title{Ingest Timor Landings Survey data}
\usage{
ingest_timor_landings(log_threshold = logger::DEBUG)
}
\arguments{
\item{log_threshold}{The (standard Apache logj4) log level used as a
threshold for the logging infrastructure. See \link[logger:log_levels]{logger::log_levels} for more
details}
}
\value{
No output. This funcrion is used for it's side effects
}
\description{
Downloads landings information that has been collected using Kobo Toolbox and
uploads it to cloud storage services.
}
\details{
This function downloads the survey metadata (survey information) as well as
the survey responses. Afterwards it uploads this information to cloud
services. To improve reproducibility, the file names used contain a
versioning string that includes the date-time and, if available, the first 7
digits of the git commit sha. This is acomplished using \code{\link[=add_version]{add_version()}}

The information about the surveys and the cloud storage is obtained from the
configuration file \code{conf.yml}. This file is located in the \emph{inst} directory
of the package and can be accessed using
\code{system.file("conf.yml",package="peskas.timor.data.pipeline")}. The specific
configuration is detemined using the environment variable \code{R_CONFIG_ACTIVE}.

Progress through the function is tracked using the package \emph{logger}.
}
\keyword{workflow}
