% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest-pds-data.R
\name{insistent_upload_cloud_file}
\alias{insistent_upload_cloud_file}
\title{Insistent version of \code{upload_cloud_file()}}
\usage{
insistent_upload_cloud_file(..., delay = 3)
}
\arguments{
\item{...}{Inputs to \code{upload_cloud_file()}}

\item{delay}{he time interval to suspend execution for, in seconds.}
}
\value{
No output. This function is used for it's side effects
}
\description{
Just like \code{upload_cloud_file()}, this function takes a vector of tracks files
as argument and upload them to the cloud. The function uses
\link[purrr:insistently]{purrr::insistently} in order to continue to upload files despite stale OAuth
token.
}
