% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge-trips.R
\name{merge_trips}
\alias{merge_trips}
\title{Merge trips}
\usage{
merge_trips()
}
\value{
no outputs. This function is used for it's side effects
}
\description{
Merges validated landings and the validated pds trips into one single data
frame. Merging is done only when there is a single landing and tracked trip
per day. Days are determined using the end of the tracking trip.
}
\details{
The parameters needed in the config file are those required for
\code{validate_landings()}, and \code{validate_pds_trips()}combined, as well as
merged_trips.
}
\keyword{workflow}
