% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-fishery.R
\name{model_value}
\alias{model_value}
\title{Model landing value}
\usage{
model_value(trips)
}
\arguments{
\item{trips}{data frame}
}
\value{
a glmmTMB model
}
\description{
Model landing value
}
\keyword{internal}
