% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peskas.timor.data.pipeline-package.R
\docType{package}
\name{peskas.timor.data.pipeline-package}
\alias{peskas.timor.data.pipeline}
\alias{peskas.timor.data.pipeline-package}
\title{peskas.timor.data.pipeline: Functions to Implement the Timor Small Scale Fisheries Data Pipeline}
\description{
This package implements the data and modelling pipelines underpinning the Peskas system.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://worldfishcenter.github.io/peskas.timor.data.pipeline/}
  \item Report bugs at \url{https://github.com/WorldFishCenter/peskas.timor.data.pipeline/issues}
}

}
\author{
\strong{Maintainer}: Lorenzo Longobardi \email{l.longobardi@cgiar.org} (\href{https://orcid.org/0000-0003-3126-7341}{ORCID})

Authors:
\itemize{
  \item Fernando Cagua \email{f.cagua@cgiar.org} (\href{https://orcid.org/0000-0001-5867-3687}{ORCID})
  \item Hamza Altarturi \email{h.altarturi@cgiar.org}
}

Other contributors:
\itemize{
  \item WorldFish [copyright holder]
}

}
\keyword{internal}
