% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess-landings.R
\name{preprocess_landings_step_1}
\alias{preprocess_landings_step_1}
\title{Pre-process Timor Landings Survey data (step 1)}
\usage{
preprocess_landings_step_1(log_threshold = logger::DEBUG)
}
\arguments{
\item{log_threshold}{The (standard Apache logj4) log level used as a threshold for the logging infrastructure. See \link[logger:log_levels]{logger::log_levels} for more details}
}
\value{
no outputs. This function is used for it's side effects
}
\description{
Downloads raw structured data from cloud storage services and pre-process
into a binary format that is easier to deal with in R.
}
\details{
In order to not exceed CPU memory limits in Docker containers, the
preprocessing of raw landings data was splitted in two containers
(two separate jobs in GitHub actions), this function process the first half
of raw data, while the function \code{preprocess_landings_step_2} process the second
half.

This function downloads the landings data from a given version (specified in
the config file \code{conf.yml}.The parameters needed are:

\if{html}{\out{<div class="sourceCode">}}\preformatted{surveys:
  landings:
    <version>:
      raw:
        file_prefix:
        version:
storage:
  google:
    key:
    options:
}\if{html}{\out{</div>}}

Progress through the function is tracked using the package \emph{logger}.
}
\keyword{workflow}
