% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess-legacy-landings.R
\name{preprocess_legacy_landings}
\alias{preprocess_legacy_landings}
\title{Pre-process Timor legacy Landings Survey data}
\usage{
preprocess_legacy_landings(log_threshold = logger::DEBUG)
}
\arguments{
\item{log_threshold}{The (standard Apache logj4) log level used as a
threshold for the logging infrastructure. See \link[logger:log_levels]{logger::log_levels} for more
details}
}
\value{
no outputs. This funcrion is used for it's side effects
}
\description{
Downloads raw structured legacy data from cloud storage services and pre-process
into a binary format that is easier to deal with in R.
}
\details{
This function downloads the landings data from a given version (specified in
the config file \code{conf.yml}. The parameters needed are:\preformatted{surveys:
  landings_legacy:
   api:
   survey_id:
   token:
   file_prefix:
 version:
  preprocess:

storage:
  storage_name:
    key:
    options:
      project:
      bucket:
      service_account_key:
}

Progress through the function is tracked using the package \emph{logger}.
}
\keyword{workflow}
