% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess-pds-trips.R
\name{preprocess_pds_tracks}
\alias{preprocess_pds_tracks}
\title{Pre-process Pelagic Data System tracks}
\usage{
preprocess_pds_tracks(log_threshold = logger::DEBUG)
}
\arguments{
\item{log_threshold}{The (standard Apache logj4) log level used as a
threshold for the logging infrastructure. See \link[logger:log_levels]{logger::log_levels} for more
details}
}
\value{
no outputs. This function is used for it's side effects
}
\description{
Downloads raw structured data from cloud storage services and pre-process
into a binary format that is easier to deal with in R.
}
\details{
The preprocessed file consists of a dataframe containing PDS tracks' diagnostics
useful to the following validation step.

The parameters needed are:

\if{html}{\out{<div class="sourceCode">}}\preformatted{pds:
 tracks:
  token: !expr Sys.getenv('PDS_TOKEN')
  secret: !expr Sys.getenv('PDS_SECRET')
  file_prefix:
pds_storage:
  google:
   options:
    bucket: pds-timor
}\if{html}{\out{</div>}}

Progress through the function is tracked using the package \emph{logger}.
}
\keyword{workflow}
