% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess-pds-trips.R
\name{preprocess_pds_trips}
\alias{preprocess_pds_trips}
\title{Pre-process Pelagic Data System trips}
\usage{
preprocess_pds_trips(log_threshold = logger::DEBUG)
}
\arguments{
\item{log_threshold}{The (standard Apache logj4) log level used as a
threshold for the logging infrastructure. See \link[logger:log_levels]{logger::log_levels} for more
details}
}
\value{
no outputs. This function is used for it's side effects
}
\description{
Downloads raw structured data from cloud storage services and pre-process
into a binary format that is easier to deal with in R.
}
\details{
This function downloads the landings data from a given version (specified in
the config file \code{conf.yml}. The parameters needed are:

\if{html}{\out{<div class="sourceCode">}}\preformatted{pds:
  trips:
    token:
    secret:
    file_prefix:
  version:
    preprocess:
storage:
  storage_name:
    key:
    options:
      project:
      bucket:
      service_account_key:
}\if{html}{\out{</div>}}

Progress through the function is tracked using the package \emph{logger}.
}
\keyword{workflow}
