% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess-metadata-tables.R
\name{pt_get_devices_table}
\alias{pt_get_devices_table}
\title{Get table containing information about the PDS devices}
\usage{
pt_get_devices_table(boats_table)
}
\arguments{
\item{boats_table}{the boat metadata table obtained from Google
Spreadsheets}
}
\value{
a data frame with columns: imei
}
\description{
Get table containing information about the PDS devices
}
