% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process_landings.R
\name{pt_nest_attachments}
\alias{pt_nest_attachments}
\title{Nest survey attachment columns in landings data}
\usage{
pt_nest_attachments(x)
}
\arguments{
\item{x}{A data frame containing landings data from the Timor operations
in which the attachments information is spread across multiple columns}
}
\value{
Landings data in which the information about multiple attachments has
been nested into a single column
}
\description{
Nest survey attachment columns in landings data
}
\author{
Fernando Cagua
}
