% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess-metadata-tables.R
\name{pt_validate_devices}
\alias{pt_validate_devices}
\title{Parse and validate devices table}
\usage{
pt_validate_devices(devices_table)
}
\arguments{
\item{devices_table}{a data frame with the devices}
}
\value{
a tibble
}
\description{
Convert date and date-time columns and ensures that device_imei is stored as
a character. Currently this table performs no validations.
}
\keyword{preprocessing}
