% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess-metadata-tables.R
\name{pt_validate_fao_catch}
\alias{pt_validate_fao_catch}
\title{Parse and validate fao catch table}
\usage{
pt_validate_fao_catch(fao_catch_table)
}
\arguments{
\item{fao_catch_table}{a data frame with the catch type info}
}
\value{
a tibble
}
\description{
Currently this function does not perform anything and is a placeholder
}
