% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-dataverse.R
\name{publish_last_dataset}
\alias{publish_last_dataset}
\title{Publish latest dataset created}
\usage{
publish_last_dataset(key = NULL, dataverse = NULL, server = NULL)
}
\arguments{
\item{key}{API token associated to the Dataverse account.}

\item{dataverse}{A character string specifying the Dataverse ID.}

\item{server}{A character string specifying a Dataverse server.}
}
\description{
The function converts from "draft" to "public" the latest dataset uploaded
in a specific Dataverse repository.
}
\examples{
\dontrun{
publish_last_dataset(key = "my_key", dataverse = "my_dataverse", server = "dataverse.example.com")
}
}
\keyword{export}
