% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve-pds-data.R
\name{retrieve_pds_tracks}
\alias{retrieve_pds_tracks}
\title{Download pelagic data system tracks}
\usage{
retrieve_pds_tracks(prefix, secret, token, id)
}
\arguments{
\item{prefix}{Name to be used as the prefix of the file names to be
downloaded. Can be a path.}

\item{secret}{Access secret code for the account}

\item{token}{Access token for the account}

\item{id}{Trip unique id from which download tracks data}
}
\value{
A character vector with paths of the downloaded files
}
\description{
Download pelagic data system tracks
}
\examples{
\dontrun{
retrieve_pds_tracks(
  prefix = "my-tracks", secret = "abcXXXXXX",
  token = "123XXXXXX", id = "2327722"
)
# To download in a different path
dir.create("my-data-dir")
retrieve_pds_tracks(
  prefix = "my-data-dir/my-tracks", secret = "abcXXXXXX",
  token = "123XXXXXX", id = "2327722"
)
}
}
