% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve-pds-data.R
\name{retrieve_pds_tracks_data}
\alias{retrieve_pds_tracks_data}
\title{Download pelagic data system tracks in as csv}
\usage{
retrieve_pds_tracks_data(
  path,
  secret = NULL,
  token = NULL,
  id = NULL,
  overwrite = TRUE
)
}
\arguments{
\item{path}{String with path to file where API request should be saved}

\item{secret}{Access secret code for the account}

\item{token}{Access token for the account}

\item{id}{Trip unique id from which download tracks data}

\item{overwrite}{Will only overwrite existing path if TRUE.}
}
\value{
The file path
}
\description{
Download pelagic data system tracks in as csv
}
\examples{
retrieve_pds_tracks_data("test.csv")
file.remove("test.csv")
}
