% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve-pds-data.R
\name{retrieve_pds_trips}
\alias{retrieve_pds_trips}
\title{Download pelagic data system trips}
\usage{
retrieve_pds_trips(
  prefix,
  secret = NULL,
  token = NULL,
  start_date = "2018-07-01",
  end_date = Sys.Date(),
  append_version = TRUE
)
}
\arguments{
\item{prefix}{Name to be used as the prefix of the file names to be
downloaded. Can be a path.}

\item{secret}{Access secret code for the account}

\item{token}{Access token for the account}

\item{start_date}{Start date from which download trips information}

\item{end_date}{Last date to download trips information}

\item{append_version}{whether to append versioning information to the
filename using \link{add_version}.}
}
\value{
A character vector with paths of the downloaded files
}
\description{
Download pelagic data system trips
}
\examples{
\dontrun{
retrieve_pds_trips(
  prefix = "my-trips", secret = "abcXXXXXX",
  token = "123XXXXXX", start_date = "2018-07-01",
  end_date = Sys.Date()
)
# To download in a different path
dir.create("my-data-dir")
retrieve_pds_trips(
  prefix = "my-data-dir/my-trips", secret = "abcXXXXXX",
  token = "123XXXXXX", start_date = "2018-07-01",
  end_date = Sys.Date()
)
}
}
