% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve-pds-data.R
\name{retrieve_pds_trips_data}
\alias{retrieve_pds_trips_data}
\title{Download pelagic data system trips in as csv or json}
\usage{
retrieve_pds_trips_data(
  path,
  secret = NULL,
  token = NULL,
  start_date = NULL,
  end_date = NULL,
  overwrite = TRUE
)
}
\arguments{
\item{path}{String with path to file where API request should be saved}

\item{secret}{Access secret code for the account}

\item{token}{Access token for the account}

\item{start_date}{Start date from which download trips information}

\item{end_date}{Last date to download trips information}

\item{overwrite}{Will only overwrite existing path if TRUE.}
}
\value{
The file path
}
\description{
Download pelagic data system trips in as csv or json
}
\examples{
retrieve_pds_trips_data("test.csv")
file.remove("test.csv")
}
