% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve-survey-data.R
\name{retrieve_survey}
\alias{retrieve_survey}
\title{Download survey}
\usage{
retrieve_survey(
  prefix,
  api,
  id,
  token,
  format = c("csv", "json"),
  metadata = TRUE,
  append_version = TRUE
)
}
\arguments{
\item{prefix}{name to be used as the prefix of the file names to be
downloaded. Can be a path.}

\item{api}{Either "kobo", "kobohr", "ona", or a custom (full) URL. API URLs
are made available for KoBo Toolbox ("kobo",
\url{https://kc.kobotoolbox.org/api/v1/}), KoBo Humanitarian Response
("kobohr", \url{https://kc.humanitarianresponse.info/api/v1/}), Ona ("ona",
\url{https://api.ona.io/api/v1/}) and Unhcr ("unhcr",
\url{https://kobocat.unhcr.org/api/v1/}) . For your own installation, or
other installations using the same API but accessed at a different URL,
enter the full URL.}

\item{id}{survey id. Usually a 6 digit number. See \href{https://support.kobotoolbox.org/pulling_data_into_powerbi.html}{this support page} for
an example on how this can be obtained}

\item{token}{access token for the account e.g. "Token XXXXXXX"}

\item{format}{Either "csv" or "json"}

\item{metadata}{whether to download metadata as well as data}

\item{append_version}{whether to append versioning information to the
filename using \link{add_version}.}
}
\value{
a character vector with paths of the downloaded files
}
\description{
Download multiple survey artefacts (data and metadata) at once. Progress
through the function is tracked using the package \emph{logger}.
}
\examples{
\dontrun{
# It's only possible to donwload survey metadata with the account Token
retrieve_survey(
  prefix = "my-survey", api = "kobohr", id = 753491,
  token = "123XXXXXX"
)
# To download in a different path
dir.create("my-data-dir")
retrieve_survey(
  prefix = "my-data-dir/my-survey", api = "kobohr", id = 753491,
  token = "123XXXXXX"
)
}
}
