% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve-survey-data.R
\name{retrieve_survey_metadata}
\alias{retrieve_survey_metadata}
\title{Download survey metadata}
\usage{
retrieve_survey_metadata(id = NULL, token = NULL, api = "kobohr")
}
\arguments{
\item{id}{survey id. Usually a 6 digit number. See \href{https://support.kobotoolbox.org/pulling_data_into_powerbi.html}{this support page} for
an example on how this can be obtained}

\item{token}{access token for the account e.g. "Token XXXXXXX"}

\item{api}{Either "kobo", "kobohr", "ona", or a custom (full) URL. API URLs
are made available for KoBo Toolbox ("kobo",
\url{https://kc.kobotoolbox.org/api/v1/}), KoBo Humanitarian Response
("kobohr", \url{https://kc.humanitarianresponse.info/api/v1/}), Ona ("ona",
\url{https://api.ona.io/api/v1/}) and Unhcr ("unhcr",
\url{https://kobocat.unhcr.org/api/v1/}) . For your own installation, or
other installations using the same API but accessed at a different URL,
enter the full URL.}
}
\value{
A list with survey metadata
}
\description{
Download survey metadata
}
\examples{
\dontrun{
# It's only possible to donwload survey metadata with the account Token
retrieve_survey_metadata(753491, token = "123XXXXXX")
}

}
