% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{send_sites_report}
\alias{send_sites_report}
\title{Send sites report email}
\usage{
send_sites_report(log_threshold = logger::DEBUG)
}
\arguments{
\item{log_threshold}{The (standard Apache logj4) log level used as a
threshold for the logging infrastructure. See \link[logger:log_levels]{logger::log_levels} for more
details}
}
\value{
Nothing, this function is useful for its side effects.
}
\description{
This function takes advantage of the package \code{blastula} to send an email
containing a summary report in each landig site.
}
