% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud-storage.R
\name{upload_cloud_file}
\alias{upload_cloud_file}
\title{Upload a local file to a cloud storage bucket}
\usage{
upload_cloud_file(file, provider, options, name = file)
}
\arguments{
\item{file}{a file-path (character) to upload. A vector with multiple files
is also supported.}

\item{provider}{cloud provider to use, either "gcs" or "aws"}

\item{options}{named list with cloud provider options, see details}

\item{name}{What to call the file once uploaded. Default is the filepath}
}
\value{
If \code{provider} is "gcs" and if successful a list of medatada objects
}
\description{
Upload a local file to a cloud storage bucket
}
\details{
\subsection{Google Cloud Services}{

For Google Cloud Services ("gcs") options must be a list with two fields:
\code{bucket} with the bucketname (character) you are uploading to, and
\code{service_account_key} with the contents of the authentication json file you
have downloaded from your Google Project (if \link{cloud_storage_authenticate} has
not been called before).

This function uses \link[googleCloudStorageR:gcs_upload]{googleCloudStorageR::gcs_upload} under the hood to upload
the file.
}
}
\examples{

# Google Cloud Services
\dontrun{
authentication_details <- readLines("location_of_json_file.json")
upload_cloud_file(
  file = "table_to_upload.csv",
  provider = "gcs",
  options = list(
    service_account_key = authentication_details,
    bucket = "my-bucket"
  )
)
}

}
