% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-dataverse.R
\name{upload_files}
\alias{upload_files}
\title{Upload files to Dataverse}
\usage{
upload_files(file_list = NULL, key = NULL, dataverse = NULL, server = NULL)
}
\arguments{
\item{file_list}{Paths indicating the files to be uploaded to Dataverse.}

\item{key}{API token associated to the Dataverse account.}

\item{dataverse}{A character string specifying the Dataverse ID.}

\item{server}{A character string specifying a Dataverse server.}
}
\description{
This function upload a list files in a specific Dataverse repository.
}
\examples{
\dontrun{
upload_files(
  file_list = c("file1.tsv", "file2.tsv"),
  key = "my_key",
  dataverse = "my_dataverse",
  server = "dataverse.example.com"
)
}
}
\keyword{export}
