% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-pds-trips.R
\name{validate_pds_navigation}
\alias{validate_pds_navigation}
\title{Validate pds trips duration and distance}
\usage{
validate_pds_navigation(data)
}
\arguments{
\item{data}{A data frame containing pds trips}
}
\value{
A list containing data frames with validated catch duration and
catch distance traveled
}
\description{
This function takes pds trips information to return validated data frames of
trip duration and distance.
}
\examples{
\dontrun{
  pars <- read_config()
  pds_trips <- retrieve_pds_trips(prefix = pars$pds$trips$file_prefix,
                                  secret = pars$pds$trips$secret,
                                  token = pars$pds$trips$token)\%>\%
 readr::read_csv(col_types = readr::cols(.default = readr::col_character()))

  validate_pds_navigation(pds_trips)
}

}
