% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions.R
\name{validate_surveys_time}
\alias{validate_surveys_time}
\title{Validate surveys' temporal parameters}
\usage{
validate_surveys_time(submissions, hrs = NULL)
}
\arguments{
\item{submissions}{Output of \code{\link[=validation_submissions]{validation_submissions()}}.}

\item{hrs}{Limit of trip duration in hours to be considered a valid catch
session.}
}
\value{
A list containing data frames with validated catch dates and catch
duration.
}
\description{
This function takes the submission-level view of the landings and validates
the temporal info associated to each survey.
}
\examples{
\dontrun{
conf <- read_config()
landings <- get_weighted_landings(conf)
validate_surveys_time(validation_submissions(landings), hrs = 18)
}
}
\keyword{validation}
