% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions.R
\name{alert_outlier}
\alias{alert_outlier}
\title{Generate an alert vector based on the \code{univOutl::LocScaleB()} function}
\usage{
alert_outlier(
  x,
  alert_if_larger,
  alert_if_smaller = alert_if_larger,
  no_alert_value = NA_real_,
  ...
)
}
\arguments{
\item{x}{numeric vector where outliers will be checked}

\item{alert_if_larger}{alert for when x is above the bounds found by \code{univOutl::LocScaleB()}}

\item{alert_if_smaller}{alert for when x is below the bounds found by \code{univOutl::LocScaleB()}}

\item{no_alert_value}{value to put in the output when there is no alert (x is within bounds)}

\item{...}{arguments for \code{univOutl::LocScaleB()}}
}
\value{
a vector of the same lenght as x
}
\description{
Generate an alert vector based on the \code{univOutl::LocScaleB()} function
}
