% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-raw-data.R
\name{clean_updated_landings}
\alias{clean_updated_landings}
\title{Clean Updated (Peskas 2) landings data}
\usage{
clean_updated_landings(x)
}
\arguments{
\item{x}{Data frame containing raw data from updated landings (Peskas 2).}
}
\value{
Data frame containing updated landings data in the same syntax as
old landings.
}
\description{
New Peskas 2 survey landings include some new questions and some new edits.
This function facilitates the integration of the new survey data with the
legacy ones.
}
\examples{
\dontrun{
authentication_details <- readLines("location_of_json_file.json")
# obtain the latest version of all files corresponding to timor-landings-v1
peskas2_data <-
  cloud_object_name(
    prefix = "timor-landings-v3",
    version = "latest",
    provider = "gcs",
    options = list(
      service_account_key = authentication_details,
      bucket = "my-bucket"
    )
  )

peskas2_raw <-
  readr::read_csv(file = peskas2_data, col_types = readr::cols(.default = readr::col_character()))

clean_updated_landings(legacy_raw)
}
}
