% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pds-maps.R
\name{convert_taxa_names}
\alias{convert_taxa_names}
\title{Assign each taxon code to a broad species group}
\usage{
convert_taxa_names(data)
}
\arguments{
\item{data}{A dataframe with taxa codes under a column named "catch_taxon"}
}
\value{
\code{data} with a \code{fish_group} column
}
\description{
Assign each taxon code to a broad species group
}
\keyword{helper}
