% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pds-maps.R
\name{convert_taxa_names}
\alias{convert_taxa_names}
\title{Convert taxa codes to common names}
\usage{
convert_taxa_names(data, conf)
}
\arguments{
\item{data}{A dataframe with taxa codes under a column named "catch_taxon"}

\item{conf}{The config file}
}
\value{
A dataframe with taxa common names
}
\description{
Convert taxa codes to common names
}
\keyword{helper}
