% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-dataverse.R
\name{delete_dataset}
\alias{delete_dataset}
\title{Delete a dataset of a dataverse collection}
\usage{
delete_dataset(key, id, server)
}
\arguments{
\item{key}{API token associated to the Dataverse account.}

\item{id}{The dataset ID.}

\item{server}{A character string specifying a Dataverse server.}
}
\description{
This function delete a specific draft dataset.
}
