% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-dataverse.R
\name{delete_dataverse}
\alias{delete_dataverse}
\title{Delete a dataverse collection}
\usage{
delete_dataverse(key, dataverse, server)
}
\arguments{
\item{key}{API token associated to the Dataverse account.}

\item{dataverse}{A character string specifying the Dataverse ID.}

\item{server}{A character string specifying a Dataverse server.}
}
\description{
This function delete a specific Dataverse repository.
}
