% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataverse.R
\name{export_files}
\alias{export_files}
\title{Upload and publish a dataset on Dataverse}
\usage{
export_files(token, dataverse, server)
}
\arguments{
\item{token}{API token associated to the Dataverse account.}

\item{dataverse}{A character string specifying the Dataverse ID.}

\item{server}{A character string specifying a Dataverse server.}
}
\description{
This function upload and publish data on a specific Dataverse repository
including associated metadata information.
}
