% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-dataverse.R
\name{generate_description}
\alias{generate_description}
\title{Generate data description}
\usage{
generate_description(...)
}
\arguments{
\item{...}{unused; for backwards compatibility only}
}
\value{
A list with the description of each variable of each uploaded dataset.
}
\description{
This function generate the description of the data useful for the README file.
}
