% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataverse.R
\name{generate_metadata}
\alias{generate_metadata}
\title{Generate a list of metadata}
\usage{
generate_metadata()
}
\value{
A list with metadata information
}
\description{
The function generate a list of metadata information to append to the
files to upload to a Dataverse repository.
}
