% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-weights.R
\name{get_catch_types}
\alias{get_catch_types}
\title{Download metadata catch types}
\usage{
get_catch_types(pars)
}
\arguments{
\item{pars}{The configuration file}
}
\value{
A dataframe containing taxonomic info and common names associated to
species groups.
}
\description{
This function downloads airtable tables containing catches metadata.
}
