% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutrients.R
\name{get_fao_composition}
\alias{get_fao_composition}
\title{Get FAO Food Composition Data}
\usage{
get_fao_composition()
}
\value{
A tibble with values for various nutrients for each category of marine
food items. The columns include \code{Protein_mu}, \code{Calcium_mu}, \code{Iron_mu}, \code{Zinc_mu},
\code{Selenium_mu}, \code{Vitamin_A_mu}, and \code{Omega_3_mu}.
}
\description{
This function retrieves and processes food composition data from the FAO database.
It specifically focuses on marine food items such as octopus, squids, cockles, shrimps,
crabs, and lobsters. The data is filtered, categorized for various nutrients
including protein, calcium, iron, zinc, selenium, vitamin A, and omega-3 fatty acids.
}
\details{
The function first reads a CSV file from a given URL using the \code{read_csv} function from
the \code{readr} package. It then defines specific codes for various marine food groups:
octopus, squids, cockles, shrimps, crabs, and lobsters. The FAO composition data is then
filtered for these food items in their raw state. A new \code{interagency_code} is created
for categorization purposes. The nutrients considered are protein, calcium,
iron, zinc, selenium, vitamin A, and omega-3 fatty acids.
}
\examples{
get_fao_composition()

}
\keyword{helper}
