% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-weights.R
\name{get_fish_length}
\alias{get_fish_length}
\title{Download species body-length parameters}
\usage{
get_fish_length(taxa, rank = NULL, country_code = NULL)
}
\arguments{
\item{taxa}{The name of the species/groups from which retrieve lengths parameters.}

\item{rank}{The taxonomical ranking, can be \code{Species}, \code{Genus}, \code{Family}, or
\code{comm_name.} The latter is an abbreviation for the  common name of a species/group.}

\item{country_code}{The code of the country associated to FishBase database, default
code include Timor-Leste and Indonesia. Full list at
\url{https://www.fishbase.se/country/ListOfCountryCodes.php}.}
}
\value{
A data frame containing species length parameters
}
\description{
This function takes species, group of species or common names as arguments and
returns a data frame containing weigth-length parameters from
FishBase database.
}
\examples{
\dontrun{
get_fish_length(
  taxa = "Sardine", rank = "comm_name",
  country_code = 626
)
}
}
