% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-cloud-files.R
\name{get_full_tracks}
\alias{get_full_tracks}
\title{Download tracks data in a single file}
\usage{
get_full_tracks(pars)
}
\arguments{
\item{pars}{The configuration file.}
}
\value{
A dataframes with tracks coordinates for each trip
}
\description{
Download the latest version of PDS single-file tracks data.
}
\keyword{storage}
