% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-cloud-files.R
\name{get_full_trips}
\alias{get_full_trips}
\title{Get trips from single-file tracks data.}
\usage{
get_full_trips(pars)
}
\arguments{
\item{pars}{the configuration file.}
}
\value{
A vector of unique trips.
}
\description{
Download the list of trips from the latest version of PDS single-file
tracks data.
}
\keyword{storage}
