% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-landings.R
\name{get_merged_landings}
\alias{get_merged_landings}
\title{Download merged landings}
\usage{
get_merged_landings(conf, suffix = "")
}
\arguments{
\item{conf}{Configuration file.}

\item{suffix}{A character indicating dataframe version. Use "_weight" to download
version with calculated catch weight.}
}
\value{
A dataframe.
}
\description{
Download validated surveys landings and PDS trips.
}
