% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-cloud-files.R
\name{get_merged_trips}
\alias{get_merged_trips}
\title{Download Peskas surveys and PDS data}
\usage{
get_merged_trips(conf, ...)
}
\arguments{
\item{conf}{The configuration file}

\item{...}{Any additional parameters, passed on to
\code{\link[coasts:cloud_object_name]{coasts::cloud_object_name()}} — for example \code{version} or \code{exact_match}.}
}
\value{
A dataframe of validated survey landings and PDS trips.
}
\description{
Download validated Peskas surveys and PDS data merged by temporal matching
}
\keyword{storage}
