% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-cloud-files.R
\name{get_merged_trips}
\alias{get_merged_trips}
\title{Download Peskas surveys and PDS data}
\usage{
get_merged_trips(pars, ...)
}
\arguments{
\item{pars}{The configuration file}

\item{...}{Any additional parameters}
}
\value{
A dataframe of validated survey landings and PDS trips.
}
\description{
Download validated Peskas surveys and PDS data merged by temporal matching
}
