% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutrients.R
\name{get_nutrients_table}
\alias{get_nutrients_table}
\title{Get taxa nutritional values}
\usage{
get_nutrients_table(conf, expanded = NULL, summarise = TRUE, convert = TRUE)
}
\arguments{
\item{conf}{The configuration file.}

\item{expanded}{Optional pre-computed taxon-to-species expansion from
\code{\link[=get_morphometric_tables]{get_morphometric_tables()}}, to avoid a second FishBase round-trip.}

\item{summarise}{Whether you want to summarise the nutritional values for each
group. The function use the median by default.}

\item{convert}{Whether you want to convert the nutritional values units to
grams. See details for the default starting units.}
}
\value{
A data frame with nutritional values for each taxa group.
}
\description{
Download the modelled nutrional properties of each taxa according to
Hicks et al., 2019 (\url{https://www.nature.com/articles/s41586-019-1592-6}).
The table of nutrients for each taxon is directly called from the paper
repository (\url{https://github.com/mamacneil/NutrientFishbase}).
}
\details{
Starting units for each nutrient:
\itemize{
\item mg/100g: Calcium, Iron and Zinc.
\item μg/100g: Selenium and Vitamin A.
\item g/100g: Omega 3 and Protein.
}
}
