% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-cloud-files.R
\name{get_preprocessed_sheets}
\alias{get_preprocessed_sheets}
\title{Download Peskas metadata}
\usage{
get_preprocessed_sheets(pars)
}
\arguments{
\item{pars}{The configuration file}
}
\description{
Download preprocessed Peskas metadata from Google Sheets
}
