% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-cloud-files.R
\name{get_sync_tracks}
\alias{get_sync_tracks}
\title{Download and synchronize tracks data in a single file}
\usage{
get_sync_tracks(pars)
}
\arguments{
\item{pars}{The configuration file.}
}
\value{
The tabulated tracks collected in Timor..
}
\description{
This function downloads, synchronize and ingest tracks data from cloud
storage in a single file. Since the transferred data is relatively large,
the synchronization and uploading are refreshed when the complete track file
is outdated by at least 5000 trips.
}
\keyword{storage}
