% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest-pds-data.R
\name{get_timor_boundaries}
\alias{get_timor_boundaries}
\title{Extract Timor Geographical Boundaries}
\usage{
get_timor_boundaries(log_threshold = logger::DEBUG)
}
\arguments{
\item{log_threshold}{The logging threshold level as defined by the \code{logger} package. This parameter
allows the user to set the verbosity of logging output. The default level is \code{logger::DEBUG}.}
}
\value{
An \code{sf} object containing the geographical boundaries of Timor's regions, including adjustments
for specific areas like Atauro and Dili. The returned object includes region names and their
corresponding geometries.
}
\description{
This function extracts the geographical boundaries of Timor, including national boundaries,
administrative regions, and special cases for Atauro and Dili. It uses shapefiles stored within
the 'peskas.timor.data.pipeline' package to perform this extraction. Logging is incorporated to
inform the user of the process status and any potential threshold-based logging details.
}
\examples{
# Assuming 'peskas.timor.data.pipeline' package is installed and loaded
timor_boundaries <- get_timor_boundaries()
plot(timor_boundaries$geometry)

}
