% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess-pds-trips.R
\name{get_tracks_descriptors}
\alias{get_tracks_descriptors}
\title{Extract summaries from pds tracks}
\usage{
get_tracks_descriptors(Trip, conf, tracks_list)
}
\arguments{
\item{Trip}{A vector of pds trips to process.}

\item{conf}{The configuration file.}

\item{tracks_list}{The list of pds tracks files.}
}
\value{
A dataframe with summaries for each pds trip ID.
}
\description{
This function extracts summaries from the tracks of pds data useful to the
validation. The function extracts \code{start_end_distance}: the distance
between the initial and the final point of a trip, the \code{outliers_proportion}:
the proportion of track points exceeding 30 m/s in a trip,
the \code{timetrace_dispersion}: a measure of the the irregularity in the tracking
signal.
}
