% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess-pds-trips.R
\name{get_tracks_descriptors}
\alias{get_tracks_descriptors}
\title{Extract summaries from pds tracks}
\usage{
get_tracks_descriptors(pars, pds_trips = NULL, pds_tracks = NULL)
}
\arguments{
\item{pars}{The configuration file.}

\item{pds_trips}{The table of pds trips.}

\item{pds_tracks}{The list of pds tracks files.}
}
\value{
A dataframe with summaries for each pds trip ID.
}
\description{
This function extracts summaries from pds tracks potentially useful to the
validation step. The function extracts the \code{start_end_distance}: the distance
between the initial and the final point of a trip, the \code{outliers_proportion}:
the proportion of track points exceeding 30 m/s, the \code{timetrace_dispersion}:
measures the irregularity in the track signal.
}
