% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-cloud-files.R
\name{get_tracks_ids}
\alias{get_tracks_ids}
\title{Get pds IDs}
\usage{
get_tracks_ids(pars)
}
\arguments{
\item{pars}{the configuration file.}
}
\description{
Get the list of pds-tracks IDs stored in the pds bucket
}
