% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-cloud-files.R
\name{get_tracks_map}
\alias{get_tracks_map}
\title{Get map of Timor pds tracks.}
\usage{
get_tracks_map(pars)
}
\arguments{
\item{pars}{the configuration file.}
}
\description{
Download the static map of Timor pds tracks.
}
