% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-cloud-files.R
\name{get_validated_landings}
\alias{get_validated_landings}
\title{Download Peskas validated landings}
\usage{
get_validated_landings(log_threshold = logger::DEBUG)
}
\arguments{
\item{pars}{The configuration file}
}
\description{
Download validated Peskas data from Google Cloud
}
