% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-cloud-files.R
\name{get_validated_landings}
\alias{get_validated_landings}
\title{Download Peskas validated landings}
\usage{
get_validated_landings(log_threshold = logger::DEBUG)
}
\arguments{
\item{log_threshold}{The (standard Apache logj4) log level used as a
threshold for the logging infrastructure. See \link[logger:log_levels]{logger::log_levels} for more
details}
}
\value{
A dataframe of validated survey landings.
}
\description{
Download validated Peskas data from Google Cloud.
}
