% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-cloud-files.R
\name{get_validation_sheet}
\alias{get_validation_sheet}
\title{Get peskas validation sheet}
\usage{
get_validation_sheet(pars)
}
\arguments{
\item{pars}{the configuration file.}
}
\description{
Get the peskas validation backup sheet from google cloud
}
