% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest-pds-data.R
\name{ingest_complete_tracks}
\alias{ingest_complete_tracks}
\title{Ingest tracks data as a single file}
\usage{
ingest_complete_tracks(conf, data = NULL, trips = NULL)
}
\arguments{
\item{conf}{The configuration file.}

\item{data}{An rds file containing tracks data.}

\item{trips}{A vector of unique Trips from the argument \code{data}.}
}
\value{
No output. This function is used for it's side effects
}
\description{
This function uploads two files: \code{data}, the complete tracks in a single rds
file and \code{trips}, a vector containing unique the trips from \code{data} useful to
take track of the synchronization status of \code{data}.
}
