% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest-pds-data.R
\name{ingest_kepler_tracks}
\alias{ingest_kepler_tracks}
\title{Ingest a Kepler.gl map}
\usage{
ingest_kepler_tracks(log_threshold = logger::DEBUG)
}
\arguments{
\item{log_threshold}{The (standard Apache logj4) log level used as a
threshold for the logging infrastructure. See \link[logger:log_levels]{logger::log_levels} for more
details.}
}
\value{
Nothing. This function upload to GCS.
}
\description{
This function use the python library Kepler.gl \url{https://docs.kepler.gl/docs/keplergl-jupyter} to generate and upload a
map of PDS tracks around Timor. It uses \link[reticulate:import]{reticulate::import_from_path} to load
and run the python script \code{kepler_mapper.py}.
}
