% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge-trips.R
\name{ingest_pds_matched_trips}
\alias{ingest_pds_matched_trips}
\title{Ingest PDS matched trip}
\usage{
ingest_pds_matched_trips(log_threshold = logger::DEBUG)
}
\arguments{
\item{log_threshold}{The (standard Apache logj4) log level used as a
threshold for the logging infrastructure. See \link[logger:log_levels]{logger::log_levels} for more
details}
}
\value{
no outputs. This function is used for it's side effects
}
\description{
Ingest matched PDS tracks and landing surveys in a compressed zip folder.
GPS tracks file are uploaded in two files: one using the original temporal
resolution (1 sec recording frequency) and one using a lower temporal
resolution based on a 1 minute scale.
}
