% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest-pds-data.R
\name{ingest_pds_tracks}
\alias{ingest_pds_tracks}
\title{Ingest Pelagic Data System tracks data}
\usage{
ingest_pds_tracks(log_threshold = logger::DEBUG)
}
\arguments{
\item{log_threshold}{The (standard Apache logj4) log level used as a
threshold for the logging infrastructure. See \link[logger:log_levels]{logger::log_levels} for more
details}
}
\value{
No output. This funcrion is used for it's side effects
}
\description{
Downloads Pelagic Data System (pds) trips information and uploads it to cloud
storage services.
}
\details{
The function  downloads and uploads only tracks data that are not yet
stored in the bucket.

The parameters needed in \code{conf.yml} are:

\if{html}{\out{<div class="sourceCode">}}\preformatted{pds:
  tracks:
    token:
    secret:
    file_prefix:
pds_storage:
  storage_name:
    key:
    options:
      project:
      bucket:
      service_account_key:
}\if{html}{\out{</div>}}

Progress through the function is tracked using the package \emph{logger}.
}
\keyword{workflow}
