% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length-params.R
\name{join_weights}
\alias{join_weights}
\title{Join length-weights info to preprocessed landings}
\usage{
join_weights(data)
}
\arguments{
\item{data}{The survey landings data frame}
}
\value{
A new landings data frame including length-weights info
}
\description{
The function integrate length-weight info from FIshBase databse to preprocessed
surveys data and convert catch labels according to the FAO nomenclature
(\url{http://www.fao.org/fishery/statistics/global-production/3/en}).
}
