% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-weights.R
\name{join_weights}
\alias{join_weights}
\title{Join length-weights and nutritional parameters info to preprocessed landings}
\usage{
join_weights(data, metadata, rfish_tab, nutrients_table)
}
\arguments{
\item{data}{The survey landings data frame}

\item{metadata}{Metadata tables}

\item{rfish_tab}{Table with length weight parameters}

\item{nutrients_table}{Table with nutritional parameters}
}
\value{
A new landings data frame including length-weights info
}
\description{
The function integrate nutritional info from the \code{get_nutrients_table} function
and length-weight info from FIshBase databse to preprocessed surveys data and
convert catch  labels according to the FAO nomenclature
(\url{http://www.fao.org/fishery/statistics/global-production/3/en}).
}
\details{
The length types used to calculate weight in fish catches include total length
(TL) in survey version 1 and fork length (FL) in survey version 2 with the
exception of the group SRX (Myliobatiformes), which uses disk width (WD).
Weights of Non-fish groups are calculated according to carapace width (CW) in
crabs, mantel length (ML) in Cephalopoda and shell length (ShL) in Bivalvia,
carapace length (CL) in lobsters.
}
