% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pds-maps.R
\name{kepler_mapper}
\alias{kepler_mapper}
\title{Generate a Kepler.gl map}
\usage{
kepler_mapper(data_path = NULL)
}
\arguments{
\item{data_path}{Data to add to map.}
}
\value{
A self-contained map in html.
}
\description{
This function is a R wrapper of \code{kepler_mapper.py}, a python script function
aimed to elaborate produce a self-contained map (in html) using the
Kepler.gl python library \url{https://docs.kepler.gl/docs/keplergl-jupyter}.
}
