% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-pds-trips.R
\name{merge_consecutive_trips}
\alias{merge_consecutive_trips}
\title{Merge short intervalled trips}
\usage{
merge_consecutive_trips(x)
}
\arguments{
\item{x}{A data frame containing raw pds trips data.}
}
\value{
A dataframe with pds trips including a new column \code{associated_to}
indicating the trips to be potentially merged.
}
\description{
The pds data are quite patchy and sometimes a single trip is recorded as 2
different trips. This function calculates the time and distance interval between the end
and the start of a trip for each boat and merge the trips that start less
than 180 minutes from the other within 2 km.
}
