% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge-landings.R
\name{merge_landings}
\alias{merge_landings}
\title{Merge recent and legacy pre-processed landings}
\usage{
merge_landings(log_threshold = logger::DEBUG)
}
\arguments{
\item{log_threshold}{The (standard Apache logj4) log level used as a
threshold for the logging infrastructure. See \link[logger:log_levels]{logger::log_levels} for more
details}
}
\value{
No outputs. This function is used for it's side effects
}
\description{
Downloads pre-processed versions of recent and legacy landings from cloud and
merge in a rds-format file using the function \link[dplyr:mutate-joins]{dplyr::full_join}.
}
\details{
The merged file is then uploaded to the cloud. The parameters needed are:

\if{html}{\out{<div class="sourceCode">}}\preformatted{surveys:
  landings:
    file_prefix:
    version:
      preprocess:
  landings_legacy:
    file_prefix:
    version:
      preprocess:
  merged_landings:
    file_prefix:
    version:
storage:
  storage_name:
    key:
    options:
      project:
      bucket:
      service_account_key:
}\if{html}{\out{</div>}}

Progress through the function is tracked using the package \emph{logger}.
}
\keyword{workflow}
