% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-catch.R
\name{model_indicators}
\alias{model_indicators}
\title{Model fisheries indicators}
\usage{
model_indicators(log_threshold = logger::DEBUG)
}
\description{
Uses the trip data to model various fisheries statistics.
}
\details{
The parameters needed in \code{conf.yml} are:

\if{html}{\out{<div class="sourceCode">}}\preformatted{models:
  file_prefix:
storage:
  storage_name:
    key:
    options:
      project:
      bucket:
      service_account_key:
}\if{html}{\out{</div>}}
}
\keyword{workflow}
