% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-fishery.R
\name{model_landings}
\alias{model_landings}
\title{Model number of landings per unit}
\usage{
model_landings(trips)
}
\arguments{
\item{trips}{data frame}
}
\value{
a glmmTMB model
}
\description{
Model number of landings per unit
}
