% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess-landings.R
\name{preprocess_landings}
\alias{preprocess_landings}
\title{Pre-process Timor Landings Survey data}
\usage{
preprocess_landings(log_threshold = logger::DEBUG)
}
\arguments{
\item{log_threshold}{The (standard Apache logj4) log level used as a
threshold for the logging infrastructure. See \link[logger:log_levels]{logger::log_levels} for more
details}
}
\value{
no outputs. This funcrion is used for it's side effects
}
\description{
Downloads raw structured data from cloud storage services and pre-process
into a binary format that is easier to deal with in R.
}
\details{
This function downloads the landings data from a given version (specified in
the config file \code{conf.yml}; this file is located in the \emph{inst} directory of
the package and can be accessed using
\code{system.file("conf.yml",package="peskas.timor.data.pipeline")}). The specific
configuration is determined using the environment variable \code{R_CONFIG_ACTIVE}.

Progress through the function is tracked using the package \emph{logger}.
}
\keyword{workflow}
