% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess-metadata-tables.R
\name{preprocess_metadata_tables}
\alias{preprocess_metadata_tables}
\title{Preprocess metadata tables}
\usage{
preprocess_metadata_tables(log_threshold = logger::DEBUG)
}
\arguments{
\item{log_threshold}{The (standard Apache logj4) log level used as a threshold for the logging infrastructure. See \link[logger:log_levels]{logger::log_levels} for more details}
}
\description{
Takes the ingested Google Sheets metadata tables, parses them and writes one
versioned preprocessed list for the rest of the pipeline to read.
}
\details{
One table remains: \code{devices}, the tracker IMEI roster read by
\code{\link[=validate_imeis]{validate_imeis()}}. Everything the Airtable frame covers is read from there,
and the small code-to-label lookups are fixed in the package.

This function requires no arguments because it retrieves the parameters from
\code{conf.yml}. The fields required are:

\if{html}{\out{<div class="sourceCode">}}\preformatted{metadata:
  spreadsheet:
    name:
  version:
    preprocess:
storage:
  storage_name:
    key:
    options:
      project:
      bucket:
      service_account_key:
}\if{html}{\out{</div>}}
}
\keyword{workflow}
