% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess-metadata-tables.R
\name{pt_validate_boats}
\alias{pt_validate_boats}
\title{Parse and validate boats table}
\usage{
pt_validate_boats(boats_table)
}
\arguments{
\item{boats_table}{a data frame with the boats info}
}
\value{
a tibble
}
\description{
Convert date and date-time columns and checks that
\itemize{
\item The recorded length of the boats are valid
}
}
