% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess-metadata-tables.R
\name{pt_validate_catch_types}
\alias{pt_validate_catch_types}
\title{Parse and validate catch types table}
\usage{
pt_validate_catch_types(catch_type_table)
}
\arguments{
\item{catch_type_table}{a data frame with the catch type info}
}
\value{
a tibble
}
\description{
Currently this function does not perform anything and is a placeholder
}
