% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess-metadata-tables.R
\name{pt_validate_centro_pescas}
\alias{pt_validate_centro_pescas}
\title{Parse and validate centro de pescas table}
\usage{
pt_validate_centro_pescas(centro_pescas_table)
}
\arguments{
\item{centro_pescas_table}{a data frame with info on centro de pescas}
}
\value{
a tibble
}
\description{
Currently this function does not perform anything and is a placeholder
}
