% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess-metadata-tables.R
\name{pt_validate_morphometric_table}
\alias{pt_validate_morphometric_table}
\title{Parse and validate morphometric table}
\usage{
pt_validate_morphometric_table(morphometric_table)
}
\arguments{
\item{morphometric_table}{a data frame with morphometric info}
}
\value{
a tibble
}
\description{
Parse and validate morphometric table
}
