% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess-metadata-tables.R
\name{pt_validate_vms_installs}
\alias{pt_validate_vms_installs}
\title{Parse and validate vms_installs table}
\usage{
pt_validate_vms_installs(vms_installs_table)
}
\arguments{
\item{vms_installs_table}{a data frame with the vms movements}
}
\value{
a tibble with the vms_installs
}
\description{
Convert date and date-time columns and checks that
\itemize{
\item The date of device installations is prior to the date of recorded damage
\item The devices are installed in a single boat
}
}
