% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-dataverse.R
\name{publish_dataverse}
\alias{publish_dataverse}
\title{Publish a Dataverse repository}
\usage{
publish_dataverse(key, dataverse, server)
}
\arguments{
\item{key}{API token associated to the Dataverse account.}

\item{dataverse}{A character string specifying the Dataverse ID.}

\item{server}{A character string specifying a Dataverse server.}
}
\description{
This function publish a specific Dataverse repository.
}
