% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_config}
\alias{read_config}
\title{Read configuration file}
\usage{
read_config()
}
\value{
the environment parameters
}
\description{
Reads configuration file in \code{conf.yml} and adds some logging lines. Wrapped
for convenience
}
