% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-weights.R
\name{retrieve_lengths}
\alias{retrieve_lengths}
\title{Retrieve weight-length parameters}
\usage{
retrieve_lengths(data, country_code)
}
\arguments{
\item{data}{A table returned from \code{get_catch_types}}

\item{country_code}{The code of the country associated to FishBase database, default
code include Timor-Leste and Indonesia. Full list at
\url{https://www.fishbase.se/country/ListOfCountryCodes.php}.}
}
\value{
A data frame containing species length parameters
}
\description{
This function takes a table returned from \code{get_catch_types} and returns detailed
weight-length parameters from each species group.
}
\examples{
\dontrun{
rank_tab <- get_catch_types(pars)
retrieve_lengths(rank_tab)
}

}
