% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{send_validation_mail}
\alias{send_validation_mail}
\title{Send validation summary email}
\usage{
send_validation_mail(log_threshold = logger::DEBUG)
}
\arguments{
\item{log_threshold}{The (standard Apache logj4) log level used as a
threshold for the logging infrastructure. See \link[logger:log_levels]{logger::log_levels} for more
details}
}
\value{
Nothing, this function is useful for its side effects.
}
\description{
This function takes advantage of the package \code{blastula} to send an email
containing a summary of the latest submissions with problems.
}
\details{
Reads the flags from the shared validation database and the alert
descriptions from \code{config.yml}'s \code{validation.alerts} block.
}
\keyword{export}
