% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-dataverse.R
\name{upload_dataverse}
\alias{upload_dataverse}
\title{Upload and publish a dataset on Dataverse}
\usage{
upload_dataverse(log_threshold = logger::DEBUG)
}
\arguments{
\item{log_threshold}{The (standard Apache logj4) log level used as a
threshold for the logging infrastructure. See \link[logger:log_levels]{logger::log_levels} for more
details}
}
\description{
This function upload and publish data on a specific Dataverse repository
including associated metadata information.
}
\keyword{export}
\keyword{workflow}
