% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest-pds-data.R
\name{upload_tracks}
\alias{upload_tracks}
\title{Upload tracks files}
\usage{
upload_tracks(file_list)
}
\arguments{
\item{file_list}{Vector of files to upload}
}
\value{
No output. This function is used for it's side effects
}
\description{
This function takes a vector of tracks files as argument and upload them to
the cloud.
}
