% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions.R
\name{validate_catch_params}
\alias{validate_catch_params}
\title{Validate surveys' catch parameters}
\usage{
validate_catch_params(data = NULL, k_ind = NULL)
}
\arguments{
\item{data}{A preprocessed data frame}

\item{k_ind}{Extension of bounds for the number of individuals}
}
\value{
A data frame containing the validated catches parameters.
}
\description{
This function takes a preprocessed landings' matrix and uses univariate
techniques (see \link[univOutl:LocScaleB]{univOutl::LocScaleB}) for the identification of outliers in
the distribution of the number of individuals per catch and their size. The
function returns a data frame with the survey id, the alert number and
a nested column \code{species_group} containing validated catches parameters.
}
\examples{
\dontrun{
conf <- read_config()
landings <- get_weighted_landings(conf)
validate_catch_params(landings, k = 3)
}

}
