% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-pds-trips.R
\name{validate_pds_data}
\alias{validate_pds_data}
\title{Validate pds trips duration and distance}
\usage{
validate_pds_data(
  data,
  max_hrs = NULL,
  min_hrs = NULL,
  km = NULL,
  se_km = NULL,
  outl = NULL,
  timet = NULL
)
}
\arguments{
\item{data}{A data frame containing pds trips}

\item{max_hrs}{Upper limit of trip duration (hours) to be considered a valid
catch session.}

\item{min_hrs}{Lower limit of trip duration (hours) to be considered a valid
catch session.}

\item{km}{Limit of trip distance traveled (Km) to be considered a valid
catch session.}

\item{se_km}{Distance between the start and end point of a trip to be
considered a valid catch session.}

\item{outl}{Limit of speed outlier points in each trip to be considered a
good quality pds track.}

\item{timet}{Limit of signal trace dispersion to be considered a
good quality pds track.}
}
\value{
A list containing data frames with validated catch duration and
catch distance traveled
}
\description{
This function takes pds trips data and returns validated
trip duration and distance.
}
\examples{
\dontrun{
pars <- read_config()
pds_trips <- get_preprocessed_trips()
validate_pds(pds_trips)
}

}
