% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-pds-trips.R
\name{validate_pds_navigation}
\alias{validate_pds_navigation}
\title{Validate pds trips duration and distance}
\usage{
validate_pds_navigation(data, hrs = NULL, km = NULL)
}
\arguments{
\item{data}{A data frame containing pds trips}

\item{hrs}{Limit of trip duration in hours to be considered a valid
catch session.}

\item{km}{Limit of trip distance traveled in Km to be considered a valid
catch session.}
}
\value{
A list containing data frames with validated catch duration and
catch distance traveled
}
\description{
This function takes pds trips data and returns validated
trip duration and distance.
}
\examples{
\dontrun{
  pars <- read_config()
  pds_trips <- get_preprocessed_trips()
  validate_pds_navigation(pds_trips)
}

}
