% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-pds-trips.R
\name{validate_pds_trips}
\alias{validate_pds_trips}
\title{Validate Pelagic Data System trips}
\usage{
validate_pds_trips(log_threshold = logger::DEBUG, hrs = 18, km = 60)
}
\arguments{
\item{hrs}{Limit of trip duration in hours to be considered a valid
catch session.}

\item{km}{Limit of trip distance traveled in Km to be considered a valid
catch session.}
}
\value{

}
\description{
Validate Pelagic Data System trips
}
