% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions.R
\name{validate_surveys_time}
\alias{validate_surveys_time}
\title{Validate surveys' temporal parameters}
\usage{
validate_surveys_time(data, hrs = NULL)
}
\arguments{
\item{data}{A preprocessed data frame}

\item{hrs}{Limit of trip duration in hours to be considered a valid
catch session.}
}
\value{
A list containing data frames with validated catch dates and
catch duration.
}
\description{
This function takes a preprocessed landings' matrix and validate temporal
info associated to each survey.
}
\examples{
\dontrun{
  pars <- read_config()
  landings <- get_merged_landings(pars)
  validate_surveys_time(landings,hrs =18)
}
}
